Add GC analysis, better file reading, better comments
The files are parsed by a better function by Joe, but I have very little idea how.  Quite hard to read.  But I tested it, and it works.
The analysis function now computes the CG content of the DNA.
I made better docstrings for the functions, to comply with PEP8, which I had never heard of before now, but I will now try hard to comply with it.